+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+# v0.3.0 (2025-03-04)
+
+* **Feature**: Improved large object performance and and fuse mount performance. Added support for buckets in multiple regions. 
+* **Dependency Update**: Updated to the latest SDK module versions.
+
 # v0.2.1 (2025-02-23)
 
-* **Feature**: Updated TLS configuration 
-* **Dependency Update**: Updated to the latest SDK module versions
+* **Feature**: Updated TLS configuration. 
+* **Dependency Update**: Updated to the latest SDK module versions.
 
 # v0.2.0 (2025-02-21)
 
@@ -11,4 +16,4 @@
 # v0.1.0 (2025-02-19)
 
 * **Feature**: First version of the SDK before launch.
-* **Dependency Update**: Updated to the latest SDK module versions
+* **Dependency Update**: Updated to the latest SDK module versions.